+++ src/ServoMotorSetting.cpp
@@ -10,7 +10,7 @@ ServoMotorSetting::~ServoMotorSetting()
 }
 
 // high time range from 1.3ms to 1.7ms: 1.3 is full speed clockwise, 1.5 is no rotation, 1.7 is full speed counterclockwise
-int ServoMotorSetting::SpeedToHighTime(int speed) // speed range from -10 to 10, 0 is stop
+int ServoMotorSetting::SpeedToHighTime(float speed) // speed range from -10 to 10, 0 is stop
 {
     if (speed > 10)
         speed = 10;
@@ -23,7 +23,7 @@ int ServoMotorSetting::SpeedToHighTime(int speed) // speed range from -10 to 10,
     return neutral_time + ((speed + 1) * step_size);
 }
 
-void ServoMotorSetting::ChangeSpeed(int speed)
+void ServoMotorSetting::ChangeSpeed(float speed)
 {
     int high_time = SpeedToHighTime(speed);
     int per = INVERSEDUTY + high_time;
